Refactor more members
