+++ src/openrct2/network/NetworkClient.h
@@ -43,7 +43,13 @@ namespace OpenRCT2
         void SendChat(const char* text);
         void SendGameAction(const GameAction* action);
 
+        const NetworkServerState_t& GetServerState() const;
+
+        bool CheckDesynchronisation();
+        bool IsDesynchronised();
+
     private:
+        bool CheckSRAND(uint32_t tick, uint32_t srand0);
         void ProcessPlayerInfo();
         void ProcessPlayerList();
 
@@ -66,6 +72,36 @@ namespace OpenRCT2
         void OnPacketObjectsList(NetworkConnection& connection, NetworkPacket& packet);
         void OnPacketScripts(NetworkConnection& connection, NetworkPacket& packet);
         void OnPacketGameState(NetworkConnection& connection, NetworkPacket& packet);
+
+    private:
+        struct PlayerListUpdate
+        {
+            std::vector<NetworkPlayer> players;
+        };
+
+        struct ServerTickData_t
+        {
+            uint32_t srand0;
+            uint32_t tick;
+            std::string spriteHash;
+        };
+
+        std::unique_ptr<NetworkConnection> _serverConnection;
+        std::map<uint32_t, PlayerListUpdate> _pendingPlayerLists;
+        std::multimap<uint32_t, NetworkPlayer> _pendingPlayerInfo;
+        std::map<uint32_t, ServerTickData_t> _serverTickData;
+        std::vector<std::string> _missingObjects;
+
+        std::string _host;
+        uint16_t _port = 0;
+
+        OpenRCT2::MemoryStream _serverGameState;
+        NetworkServerState_t _serverState;
+        uint32_t _lastSentHeartbeat = 0;
+        uint32_t server_connect_time = 0;
+        SocketStatus _lastConnectStatus = SocketStatus::Closed;
+        bool _requireReconnect = false;
+        bool _clientMapLoaded = false;
     };
 
 } // namespace OpenRCT2
